BobaSpot/BobaSpot/backend/requirements.txt
django==3.1.3
djangorestframework==3.14.0
psycopg2-binary
python_dotenv==1.0.0
jwt==1.3.1
djangorestframework-jwt==1.11.0